src/notifier/requirements.txt
psycopg2-binary==2.9.*
prometheus-client==0.20.*
python-dotenv==1.0.*
